% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedTRB.R
\docType{data}
\name{publishedTRB}
\alias{publishedTRB}
\title{Published T cell receptor beta CDR3 amino acid sequences with known
antigen specificity}
\format{A dataframe with 9,606 rows and 6 columns corresponding amino acid
sequence, PubMed ID (PMID), HLA type, antigen specificity, epitope, and
\% prevalence that the sequence appeared within the peripheral blood of 55
healthy individuals.}
\source{
Antigen specific T cell receptor beta sequences were manual curated
from published studies with help from Edus H. Warren, Andrea Towlerton,
Steve House, and Will DeWitt.  Approximatley 5900 sequences were downloaded
from VDJserver (https://vdjdb.cdr3.net).
}
\usage{
publishedTRB
}
\description{
A database of unique productive T cell receptor beta CDR3 amino acid sequences
with known antigen specificity.
}
\keyword{datasets}
