% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publishedTRB.R
\docType{data}
\name{publishedTRB}
\alias{publishedTRB}
\title{Published T cell receptor beta CDR3 amino acid sequences with known
antigen specificity}
\format{A dataframe with 3,706 rows and 6 columns corresponding amino acid
sequence, PubMed ID (PMID), HLA type, antigen specificity, epitope, and
\% prevalence that the sequence appeared within the peripheral blood of 55
healthy individuals.}
\source{
Citation for each sequence is provided as a PubMed ID (PMID) within
the database.  Steve House and Will DeWitt from Adaptive Biotechnologies
helped to curate the list of published sequences.
}
\usage{
publishedTRB
}
\description{
A database of unique productive T cell receptor beta CDR3 amino acid sequences
with known antigen specificity.
}
\keyword{datasets}

